Fixed acceptance of spaces and special characters in XML path
The parser now accepts spaces and special characters in the filename of the path.
However it encodes a binary file without these, so the binary file will have a different name than the original XML file, if it contains spaces or special characters.
The binary filename won't contain the original spaces and special characters.